bobbit: by default, don't prefix messages with nicks
To allow !give to work, have it wrap the nick string with a custom
object with a prefix attribute.  Check for this attribute in order to
prefix formatted messages.

We may wish to move the custom class out of modules/give later to allow
other modules the ability to prefix on-demand as well.