feeds: add support for a nonce field to workaround stale cache.
The Week uses Cloudflare to cache its RSS feed.  Unfortunately, it
appears that they have something misconfigured since it regularly serves
stale information.  We can workaround this by adding ?{nonce} to the end
of the URL (where nonce is a UUID), which forces Cloudflare to perform a
fresh fetch.

Ugly, but hopefully it works and is only needed for certain websites.