title: fix urls in CTCP ACTION messages
Messages sent as CTCP ACTIONs (aka with "/me") are typically surrounded
with the ASCII 001 character. The named capture group for urls in the
title command's regex pattern does not account for this fact. Therefore
if an ACTION message ends with a url, the ASCII 001 character will be
considered part of the url. For most websites this will result in a 404
and thus a non-existent or incorrect title.

Before this patch, we have the following incorrect behavior:

<deckard> i am reading https://en.wikipedia.org/wiki/Client-to-client_protocol
<bobbit>  Title: Client-to-client protocol - Wikipedia
        * deckard is reading https://en.wikipedia.org/wiki/Client-to-client_protocol
<bobbit> Title: Bad title - Wikipedia

This patch should fix this behavior by stripping the ASCII 001 character
from the end of urls. The relevant CTCP ACTION documentation is here:
https://datatracker.ietf.org/doc/html/draft-oakley-irc-ctcp-02#appendix-A.1